examples/monorepo/output/packages/shared/pyproject.toml
[project]
name = "shared-auth"
version = "0.1.0"
description = "Shared authentication middleware for the monorepo"
requires-python = ">=3.11"
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
